requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
SQLAlchemy==1.4.44
flask-sqlalchemy==3.0.2
celery==5.2.7
Flask-Cors==3.0.10
Werkzeug==2.2.2
requests==2.28.1
DateTime==4.7
python-dateutil==2.8.2
redis==4.3.5
PyJWT==2.6.0
random-username
pandas==1.5.3
pytest==7.2.2
algoliasearch==3.0.0
python-dotenv==1.0.0